+++ ts/src/program/event.ts
@@ -44,8 +44,6 @@ export class EventParser {
       }
       if (didPop) {
         execution.pop();
-        // Skip the "success" log, which always follows the consumed log.
-        logScanner.next();
       }
       log = logScanner.next();
     }
@@ -89,18 +87,22 @@ export class EventParser {
   private handleSystemLog(log: string): [string | null, boolean] {
     // System component.
     const logStart = log.split(":")[0];
+
+    // Did the program finish executing?
+    if (
+      logStart.match(/^Program (.*) success/g) !== null
+    ) {
+      return [null, true];
     // Recursive call.
-    if (logStart.startsWith(`Program ${this.programId.toString()} invoke`)) {
+    } else if (
+      logStart.startsWith(`Program ${this.programId.toString()} invoke`)
+    ) {
       return [this.programId.toString(), false];
     }
-    // Cpi call.
+    // CPI call.
     else if (logStart.includes("invoke")) {
       return ["cpi", false]; // Any string will do.
     } else {
-      // Did the program finish executing?
-      if (logStart.match(/^Program (.*) consumed .*$/g) !== null) {
-        return [null, true];
-      }
       return [null, false];
     }
   }