+++ CHANGELOG.md
@@ -11,6 +11,15 @@ incremented for features.
 
 ## [Unreleased]
 
+### Features
+
+* lang: Adds `require` macro for specifying assertions that return error codes on failure ([#483](https://github.com/project-serum/anchor/pull/483)).
+* lang: Allow one to specify arbitrary programs as the owner when creating PDA ([#483](https://github.com/project-serum/anchor/pull/483)).
+
+### Fixes
+
+* lang: Preserve all instruction data for fallback functions ([#483](https://github.com/project-serum/anchor/pull/483)).
+
 ## [0.11.0] - 2021-07-03
 
 ### Features